feat: Implement authorization gates

+++ app/Http/Controllers/EventController.php
@@ -8,6 +8,8 @@ use App\Http\Resources\EventResource;
 
 use App\Http\Traits\CanLoadRelationships;
 
+use Illuminate\Support\Facades\Gate;
+
 class EventController extends Controller
 {
     use CanLoadRelationships;
@@ -45,7 +47,7 @@ class EventController extends Controller
                 'start_time' => 'required|date',
                 'end_time' => 'required|date|after:start_time',
             ]),
-            ['user_id' => 1]
+            ['user_id' => $request->user()->id]
         ));
 
         return new EventResource($this->loadRelationships($event));
@@ -64,6 +66,8 @@ class EventController extends Controller
      */
     public function update(Request $request, Event $event)
     {
+        Gate::authorize('update-event', $event);
+
         $validated = $request->validate([
             'name' => 'sometimes|string|max:255',
             'description' => 'nullable|string',
@@ -81,7 +85,8 @@ class EventController extends Controller
      */
     public function destroy(Event $event)
     {
-        //
+        Gate::authorize('delete-event', $event);
+
         $event->delete();
         return response(status: 204);
     }
